Add list event api

+++ src/main/resources/db/migration/V20220928_005831__DDL_Adding_scoring_for_talent.sql
@@ -1,3 +1,3 @@
 ALTER TABLE talent
-ADD COLUMN final_score DOUBLE,
+ADD COLUMN final_score DOUBLE PRECISION,
 ADD COLUMN score_system JSONB;
